* Added support for java Swing widget toolkit, widgets: button,   check box, combo box, label, progress bar, radio box, slider,   spinner, table, text area, text field, toggle button, tree. * Changed real widget mapping: substituted eval of widget string names   with dictionary of abstract widget classes keyed by real widget classes. * Added check for executed avc_init in avc_connect. * Added dual install capability to setup to cope with both python and jython.

+++ setup.py
@@ -28,13 +28,11 @@
 # .-
 
 from distutils.core import setup
-import glob
-import os
-import os.path
-import re
-import string
+from shutil import copytree, rmtree
+import os, os.path
 import sys
 
+# parameters
 classifiers = """\
 Development Status :: 4 - Beta
 Intended Audience :: Developers
@@ -45,6 +43,24 @@ Topic :: Software Development :: Libraries :: Python Modules
 Operating System :: POSIX :: Linux
 """
 
+# when command is 'install', we need to differentiate among python or jython
+# install, selecting only files required respectively by python or jython.
+# When command is 'sdist' there is no difference all files all copied into src.
+if sys.argv[1] == 'install':
+  # if command is 'install', create a temporary installation dir
+  PACKAGE_DIR = {'avc':'src.setup'}
+  copytree('src','src.setup')
+  # remove files not required respectively by python or jython
+  if 'python' in os.path.split(sys.executable)[1]:
+    os.remove('src.setup/avcswing.py')
+  else:
+    map(os.remove,['src.setup/avcgtk.py','src.setup/avcqt3.py',
+      'src.setup/avcqt4.py','src.setup/avctk.py','src.setup/avcwx.py'])
+else:
+  # for all other commands use original src dir
+  PACKAGE_DIR = {'avc':'src'}
+
+# workaround older python versions
 if sys.version_info < (2, 3):
   _setup = setup
   def setup(**kwargs):
@@ -52,14 +68,16 @@ if sys.version_info < (2, 3):
       del kwargs["classifiers"]
     _setup(**kwargs)
 
+# do setup
 setup (
   name = 'avc',
-  version = '0.7.1',
+  version = '0.8.0',
   author = 'Fabrizio Pollastri',
   author_email = 'f.pollastri@inrim.it',
   maintainer = 'Fabrizio Pollastri',
   maintainer_email = 'f.pollastri@inrim.it',
   url = 'http://avc.inrim.it',
+  download_url = 'http://avc.inrim.it/dist/avc-0.8.0.tar.gz',
   license = 'http://www.gnu.org/licenses/gpl.txt',
   platforms = ['Linux'],
   description = """
@@ -68,13 +86,13 @@ setup (
   widgets and the variables of an application using the GUI.""",
   classifiers = filter(None, classifiers.split("\n")),
   long_description =  "see html/index.html or http://avc.inrim.it/",
-  package_dir = {'avc':'src'},
+  package_dir = PACKAGE_DIR,
   packages = ['avc'])
 
 # cleanup
-try:
+if os.access('MANIFEST',os.F_OK):
   os.remove('MANIFEST')
-except:
-  pass
+if os.access('src.setup',os.F_OK):
+  rmtree('src.setup')
 
 #### END
